add confirm step inline too

+++ yata/static/perso/js/targets/targets.js
@@ -29,14 +29,17 @@ $(document).on('click', 'a.target-list-faction', function(e){
 // delete target from target list button
 $(document).on('click', 'a.target-list-delete', function(e){
     e.preventDefault();
-    var targetId = $(this).attr("id").split("-").pop();
-    var reload = $("#target-list-refresh-"+targetId);
-    reload.load( "/target/target/", {
-        targetId: targetId,
-        type: "delete",
-        csrfmiddlewaretoken: getCookie("csrftoken")
-    });
-    reload.remove();
+
+    if (confirm("Are you sure you want to delete your list?")) {
+        var targetId = $(this).attr("id").split("-").pop();
+        var reload = $("#target-list-refresh-"+targetId);
+        reload.load( "/target/target/", {
+            targetId: targetId,
+            type: "delete",
+            csrfmiddlewaretoken: getCookie("csrftoken")
+        });
+        reload.remove();
+    }
 });
 
 // edit note
